requirements.txt
feedparser==6.0.11
schedule==1.2.1
python-dotenv==1.0.1
flask